Make gestures synchronous
Before gestures were read and added to a queue for later processing. Now
they will be sent a registered handler which can do what ever it wants
like add it to a queue or use it to run a callback hook.

As a side effect, double tap events are not longer detected by default
as only filtering for selected events is done before passing it off to
the handler.