Change input of readTouchEvent
gestures-libinput-writer now writes data atomically through the new
writeTouchEvent helper. The structure of the data has changed slightly.
Now instead of writing 2 chars to indicate the size of the sys and dev
names, only the total length is sent and it is now always sent event for
regular events. In other words, the input is not compatible with prior
versions.