Fix use-after-free
The event handler should free the gestureEvent when it is done, but we
incorrectly tried to reuse it to generate the transformed event
afterward.

Now we generally the transformed event before we pass it into
the gestureEventHandler.

Also changed the default gestureEventHandler to dumpAndFreeGesture so we
don't leak memory by default.