Use sample-reader when SGESTURES_HOME doesn't exist by default
Now if a user doesn't have SGESTURES_HOME, we'll use the new default
sgestures executable instead of creating the directory and coping the
sample reader into it. The default executable is based on the
sample-reader so the default behavior won't change. But this will make
the sgestures.sh a thin wrapper for people who don't care about
recompilation.

Additionally, one can now add a config.c to this repo and the default
executable will use that instead of the sample-reader.